% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collatzR.R
\name{collatz}
\alias{collatz}
\title{Perform the Collatz operation on a number}
\usage{
collatz(n)
}
\arguments{
\item{n}{A numeric value.}
}
\value{
The result of the Collatz operation.
}
\description{
Perform the Collatz operation on a number
}
