% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collatzR.R
\name{collatz_sequence}
\alias{collatz_sequence}
\title{Generate a Collatz sequence for a given starting number}
\usage{
collatz_sequence(start)
}
\arguments{
\item{start}{The starting number for the sequence.}
}
\value{
A list with the following components:
\itemize{
\item 'sequence': A numeric vector representing the Collatz sequence.
\item 'steps': The total number of steps to reach 1.
\item 'max_value': The maximum value in the Collatz sequence.
}
}
\description{
Generate a Collatz sequence for a given starting number
}
