% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collatzR.R
\name{is.even}
\alias{is.even}
\title{Determine if a number is even}
\usage{
is.even(x)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
TRUE if the number is even, FALSE otherwise.
}
\description{
Determine if a number is even
}
