% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collatzR.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Check if a number is a whole number (integer)}
\usage{
is.wholenumber(x, tolerance = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
TRUE if the number is a whole number, FALSE otherwise.
}
\description{
Check if a number is a whole number (integer)
}
