% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collatzR.R
\name{max_collatz}
\alias{max_collatz}
\title{Get the maximum value in a Collatz sequence}
\usage{
max_collatz(start)
}
\arguments{
\item{start}{The starting number for the sequence.}
}
\value{
The maximum value in the Collatz sequence.
}
\description{
Get the maximum value in a Collatz sequence
}
