% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collatzR.R
\name{visualize_collatz}
\alias{visualize_collatz}
\title{Visualize a Collatz sequence}
\usage{
visualize_collatz(start)
}
\arguments{
\item{start}{The starting number for the sequence.}
}
\description{
Visualize a Collatz sequence
}
